Time: 21 ms (57.81%), Space: 56.2 MB (29.17%) - LeetHub

+++ 0div-idbig-omega-company-tagsdiv-idbig-omega-topbardiv-classcompanytagscontainer-styleoverflow-x-scroll-flex-wrap-nowrap-div-classcompanytagscontainer-tagno-companies-found-for-this-problem-div-divdiv-classcompanytagscontainer-chevrondivsvg-version11-idicon-xmlnshttp-wwww3org-2000-svg-xmlns-xlinkhttp-wwww3org-1999-xlink-x0px-y0px-viewbox0-0-32-32-fill4087f1-xml-spacepreserve-stylewidth-20px-polygon-points16-22-6-12-74-106-16-192-246-106-26-12-polygonrect-id-x3c-transparent-rectangle-x3e-classst0-fillnone-width32-height32-rect-svg-div-div-div-div1921-eliminate-maximum-number-of-monsters/0div-idbig-omega-company-tagsdiv-idbig-omega-topbardiv-classcompanytagscontainer-styleoverflow-x-scroll-flex-wrap-nowrap-div-classcompanytagscontainer-tagno-companies-found-for-this-problem-div-divdiv-classcompanytagscontainer-chevrondivsvg-version11-idicon-xmlnshttp-wwww3org-2000-svg-xmlns-xlinkhttp-wwww3org-1999-xlink-x0px-y0px-viewbox0-0-32-32-fill4087f1-xml-spacepreserve-stylewidth-20px-polygon-points16-22-6-12-74-106-16-192-246-106-26-12-polygonrect-id-x3c-transparent-rectangle-x3e-classst0-fillnone-width32-height32-rect-svg-div-div-div-div1921-eliminate-maximum-number-of-monsters.java
@@ -1,15 +1,15 @@
 class Solution {
     public int eliminateMaximum(int[] dist, int[] speed) {
         int n = dist.length;
-        float time[] = new float[n];
+        double time[] = new double[n];
         for(int i=0; i<n; i++) {
             if(speed[i]==0) {
                 return 0;
             }
-            time[i] = (float)dist[i]/(float)speed[i];
-            
+            time[i] = (double)dist[i]/(double)speed[i];
         }
        Arrays.sort(time);
+        
         int count = 0;
        for(int i=0; i<n; i++) {
            if(time[i]-i>0) {
@@ -17,9 +17,7 @@ class Solution {
            } else {
                break;
            }
-           
        }
         return count;
-        
     }
 }
